feat: terminate all data set rails (storange and CDN) atomically (#93)
Clients & SPs should now call `terminateDataSetPayment` on the service
contract to terminate the payment rails for their dataset atomically.
The service contract will revert the `railTerminated` callback from the
Payments contract if the rail terminator is not the service contract
itself.

Piece removal is allowed after the payment is terminated for a dataset
but piece addition is not as increasing the rate of a terminated rail is
not allowed by the payments contract.

Once the dataset is beyond it's payment end epoch, all operations other
than ownership change and dataset deletion are reverted.