feat: remove service provider registry as we're moving it to it's own contract (#135)
This PR does the first part of #119 i.e. removes all the code and tests
for the service provider registry from this service contract. We will
now build a new more extensible service registry contract as specified
in #119 and upgrade the warm storage service contract to use the
registry. More PRs incoming for that.

But we can merge this standalone PR without waiting on them.