feat: semi-incremental abi generation (#188)
Building on discussion in
https://github.com/FilOzone/filecoin-services/pull/180#discussion_r2308191369
about the current `update-abi` target.

There's a couple of compromises here over the pure pattern target
approach, because:

1. If we rely on `forge inspect` we get a compile-all each time we call
it, so it's just slow and tedious; we could go with that, but we have
the build output anyway so we can just rely on that and get the same
output plus pull in build to our dependency chain
2. I can't just do `abi/%.abi.json: out/%.sol/%.json` to rely on those
because if they don't exist then they're not included in our pattern at
parse time. This could work _if_ we wanted to dump all of the compiled
contracts into abi/, but there's a _lot_ of contracts to dump in there.
So we just have a whitelist and a macro to strictly specify that we have
a rule per output file we want.

I also tried out a variant of this that put in some of our dependencies
to capture their ABI at the current commit, but it was more complex so
I'll leave that can of worms alone and we'll just focus on the contracts
in here.