feat!(ServiceProviderRegistry): delete removed product data (#295)
Reviewer @rvagg @zenground0
Removed product data is not currently being deleted.
That breaks one of our assumptions in synapse where we are using the
serviceURL field returned from getPDPService to detect if the product is
null.
Really we should be using isActive but this is how I noticed the issue.
We want to clean up this data from the registry.
#### Changes
* add failing test
* delete removed product data