feat: use random nonces for AddPieces (#317)
Allows for non-sequential piece additions. PDPVerifier allocates its own
pieceIds according to order added, we receive that as firstAdded and use
it to refer to the same pieceIds as PDPVerifier. Now, instead of using
firstAdded as a nonce, we allow for random nonces that the user now
supplies separately in extraData.

* Adds a new "nonce" into extraData
* Repurposes clientDataSetIds as clientNonces
* Moves signature to the end of extraData to be consistent with
CreateDataSet

Closes: https://github.com/FilOzone/filecoin-services/issues/305