feat: floor price set to the 0.06 USDFC/month (#320)
_Edit: This got switched to having 0.06 USDFC as the floor, not the
bytes; so now the bytes works out as ~24.576 GiB and we get to talk
about it in terms of cents per month._

(PR on top of #316, that should be merged first)

Amounts below this pay as if they are storing 24 GiB, which is
0.05859375 USDFC / month.

Closes: https://github.com/FilOzone/filecoin-services/issues/319

---

*  24 GiB = 24/1024 TiB = 0.0234375 TiB
* At 2.5 USDFC/TiB/month, price = 0.0234375 * 2.5 = 0.05859375
USDFC/month

Are we OK with 0.05859375? I have an option here of either setting 24GiB
as the floor _size_ or 0.06 as the floor _price_.

If we go with 0.06 as the floor then we're working with 24.576 GiB floor
size.

---

I also had to change the mock token from 6 decimal places to 18, cause
there's already a floor rounding bit of code in here and it was getting
stuck on a higher floor due to not enough decimal places, so everything
came out wrong.