feat(IValidator): Allow Forward Progress Despite Missing Proofs (#321)
Reviewer @rvagg
It is only important to block forward progress through the current
proving period.
Currently we don't allow forward progress through unsettled periods
without proofs.
This could be a problem if there was a rail unproven for a long time but
proofs were later resumed.
Though much less likely after #258 and #267, we would not want it to be
possible for the gap to exceed the block gas limit.

Also important to recall that the `fromEpoch` is an exclusive index
while `toEpoch` is an inclusive index.
#### Changes
* allow to settleUpTo previous period when unproven
* define settledUpTo in previously irrelevant case